moved namespaces, moved files, removed dead using
