+++ msbuild/Xamarin.MacDev.Tasks/Tasks/Rewriter.cs
@@ -1,10 +1,14 @@
 using System;
+using System.IO;
+using System.Collections.Generic;
 using System.Linq;
 using Mono.Cecil;
 using Mono.Cecil.Cil;
-using Mono.Cecil.Rocks;
+using ClassRedirector;
 
-namespace ClassRedirector {
+#nullable enable
+
+namespace Xamarin.MacDev.Tasks {
 	public class Rewriter {
 		const string runtimeName = "ObjCRuntime.Runtime";
 		const string classHandleName = "ObjCRuntime.Runtime/ClassHandles";
@@ -60,7 +64,9 @@ namespace ClassRedirector {
 			if (nativeHandleOpImplicit is null)
 				throw new Exception ($"Unable to find implicit cast in {nativeHandleName}");
 
-			foreach (var (csName, nameIndex) in map) {
+			foreach (var nameIndexPair in map) {
+				var csName = nameIndexPair.Key;
+				var nameIndex = nameIndexPair.Value;
 				var fieldDef = AddPublicStaticField (classHandles, nameIndex.ObjCName, nativeHandle);
 				AddInitializer (nativeHandleOpImplicit, processor, mtClassMapDef, nameIndex.MapIndex, fieldDef);
 				classMap [csName] = fieldDef;
@@ -247,8 +253,8 @@ namespace ClassRedirector {
 			if (index < 0)
 				return false;
 			var instr = il.Body.Instructions [index]!;
-			var operand = instr.Operand?.ToString () ?? "";
-			return instr.OpCode == OpCodes.Call && operand.Contains ("Class::GetHandle", StringComparison.Ordinal);
+			var operand = instr.Operand?.ToString () ?? string.Empty;
+			return instr.OpCode == OpCodes.Call && operand.Contains ("Class::GetHandle");
 		}
 
 		bool IsLdStr (ILProcessor il, int index)
